WIP for creating LOA based policies

+++ pdp-server/src/main/java/pdp/ip/IPInfo.java
@@ -9,7 +9,7 @@ public class IPInfo {
 
     private String  networkAddress;
     private String broadcastAddress;
-    private long capacity;
+    private double capacity;
     private boolean ipv4;
 
 }
